Add files via upload
# Amazon Clone

Welcome to my first project - an Amazon Clone! 🚀  
This project is built using **HTML** and **CSS** and replicates the basic structure and design of Amazon's homepage. It showcases my foundational skills in web development and responsive design. While it doesn't include backend functionality, it serves as a great starting point to enhance my front-end skills.  

Feel free to explore and share your feedback! 😊